Swapped GPI pins of LB1 & LB2
As a fast workaround to the problem of fast toggling lightbarrier1, the
respective GPI pins of lightbarrier1 & 2 have been swapped. Now lb2 toggles
more, but is a smaller problem, because stones are created through lb1.
This changes requires changes on the hardware wiring:
lb1 has to be connected with wires blue and green
lb2 has to be connected with wires yellow and orange