Revert last commit for regular HW-Setup
This commit is intended to be used, after the hardware has been fixed and
the potential crosstalk problem between pulse counter and lightbarrier 1
has been resolved.
At that point the wiring can be set back to original state and this commit
can be used.
lb1 has to be connected with wires yellow and orange
lb2 has to be connected with wires blue and green